requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
aiohttp
pydantic
python-multipart

